pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyswatcal"
version = "0.1.0"
description = "Python SWAT Calibration & Analysis Tool"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "PySWATCal Development Team", email = "contact@pyswatcal.org"}
]
keywords = ["swat", "calibration", "hydrology", "watershed", "modeling"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Hydrology",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-optimize>=0.9.0",
    "SALib>=1.4.7",
    "joblib>=1.3.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "tables>=3.8.0",
    "pyarrow>=12.0.0",
    "plotly>=5.14.0",
    "matplotlib>=3.7.0",
    "streamlit>=1.28.0",
    "tqdm>=4.65.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.2.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/PySWATCal"
Documentation = "https://pyswatcal.readthedocs.io"
Repository = "https://github.com/yourusername/PySWATCal"
Issues = "https://github.com/yourusername/PySWATCal/issues"

[tool.setuptools]
packages = ["pyswatcal"]

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=pyswatcal --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["pyswatcal"]
omit = ["*/tests/*", "*/test_*.py"]
